Create C++ Program to Find Quotient and Remainder.cpp
The given snippet is a C++ program that calculates the quotient and remainder of a division operation. Here's a breakdown of the code:

The program starts with the inclusion of the input/output stream library.
The main function is defined.
Inside the main function, variables for divisor, dividend, quotient, and remainder are declared.
The user is prompted to enter the value of the dividend using the "Enter dividend: " message.
The input for the dividend is read and stored in the variable dividend.
The user is prompted to enter the value of the divisor using the "Enter divisor: " message.
The input for the divisor is read and stored in the variable divisor.
The quotient is calculated by dividing dividend by divisor and stored in the variable quotient.
The remainder is calculated using the modulus operator (%) and stored in the variable remainder.
The calculated quotient and remainder are then displayed to the user using the "Quotient = " and "Remainder = " messages, respectively.
Finally, the main function returns 0, indicating successful program execution.